Easier implementation of map.py
By using an XML file with the characters, exits and items, I can import
everything easier. Exits only need to be added once, it will then find
and fill the missing exits with opposite directions and the original
room it came from. Items will be specific to that room too so they
don’t have to be made before hand