Better implementation of items.py
I’ve used XML again to optimise the approach for adding items. Items
are also added to inventory automatically and true and false strings
are converted into booleans. The list_of_items function also uses ‘an’
for vowels and ‘some’ if the name_type is in a group etc. Most of the
functions with items have been moved to readitems.py to increase
readability with main.py